refactor(core): extract renderer layout engine
